Update patch set 1
Patch Set 1:

> i didn't realize we had working DNSSEC code, so i'm surprised it was this easy.

This change is really trivial.  It's just setting a bit (or two) to politely ask the resolver to check DNSSEC for us.  We don't check the results ourselves.

In the official terminology, this change would make us a "non-validating security-aware stub resolver".  Implementing validation is the hard part, but this change is a prerequisite.

Patch-set: 1
Reviewer: Ben Schwartz <1161323@85c56323-6fa9-3386-8a01-6480fb634889>