Update patch set 2
Patch Set 2:

(3 comments)

i think most of the naming conventions (which i plan on following as slavishly as possible for consistency) are obvious/well-established, but i've commented on a few of the more interesting ones.

i also talked to jmgao verbally about whether it was worth distinguishing socket fds specifically. we decided it's "obvious" and that there's not much value in explicitly stating that you call accept on an fd that must be a socket (because "duh").

Patch-set: 2
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0