Update patch set 1
Patch Set 1:

How do you know that all kernels preserve these registers?

For example, did you test this on a variety of kernels, or have you examined the source code of the different kernels (like 3.x, 4.x) to verify this?

This seems a bit dangerous since if it changes, we might get some random crashes.

Also, would it be possible to add a test to verify that these registers are preserved, so that if something does change we'll catch the failure right away?

Patch-set: 1
CC: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>