Update patch set 1
Patch Set 1:

This is meant to be a bit of a feeler in terms of whether you'd want this functionality. If the overall approach looks good, then I'll work on converting it to use lz4 and do some on-device tests.

Any guidance as to how to test on-device would be appreciated. I'm guessing I need to replace /system/bin/linker, but how do you write to this partition?

Patch-set: 1
Reviewer: Andrew Grieve <1230912@85c56323-6fa9-3386-8a01-6480fb634889>