Update patch set 2
Patch Set 2: Presubmit-Verified+1

=== Finished presubmit run: L61100000107370118 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/497893?revision=2




TreeHugger finished with: 13 passed.







WARNING: The tip-of-tree(4368455) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4368206), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.







WARNING: The tip-of-tree(4368455) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4368183), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4368455) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4368070), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




You rock!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 2
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Treehugger Robot <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0
Label: Presubmit-Verified=+1