Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> wouldn't it make more sense to just test with a different symbol? we don't actually care about `close`, and we don't have any particular opinion on overriding `close` (as long as it passes the CTS tests that actually are about `close`)...
> 
> now this test contains more lines working around something we don't care about than testing the actually interesting thing, which makes the test unreadable. and for no benefit to us.

Yes, s/close/<API that isn't famous> was my original plan. But as long as we don't clear LD_PRELOAD, there is a small chance that the symbol that we choose can be overridden by the LD_PRELOAD'ed lib and I thought you won't be happy with it. Anyway, I am okay with changing the symbol. Any recommendation for the new symbol?

Patch-set: 1
Reviewer: Jiyong Park <1132673@85c56323-6fa9-3386-8a01-6480fb634889>