Update patch set 1
Patch Set 1:

> True, but my patchset heading off to kernel.org for aarch64's
 > vdso32 requires a complete(...) vdso unification to get approval by
 > the community, and thus I will need to add __vdso_time to be on-par
 > with an overlap of existing implementations in other architectures.
 > The addition of __vdso_time is a private one right one, but the
 > next one to be immediately following the public '[PATCH v5 12/12]
 > lib: vdso: do not expose gettimeofday, if no arch supported timer'
 > that hopefully will get approved soon.
 > 
 > You would be amazed at just how simple it is to add __vdso_time for
 > arm vdso, arm64 vdso64 and arm64 vdso32.

ah, makes sense.

 > I am submitting this because of some Qs
 > 
 > Q:
 > 1) do you want clock_getres() and time() split up?

my usual rule of thumb for vdso changes is "must benefit current devices at least".

given that, i'm happy to take clock_getres right now, because although literally no one calls it, it is in at least the upstream arm64 kernels (though i'd love to know what ARM was thinking...).

the time change doesn't meet the bar, though, so if you want to get clock_getres in now, yeah, split them up.

 > 2) is it acceptable to adjust libc/upstream-openbsd/lib/libc/gen/time.c?
 > The fact it is 'upstream' scares me.

yeah, in general the rule is "no", but in this case i'm happy to clean up afterwards. but not until a vdso `time` actually exists in a kernel that someone will ship with P. then i'll scratch my beard and think long and hard about whether it's time for me to switch to ifuncs, or whether your unification work actually means the opposite, and we've taken our first step down the path to _requiring_ that these functions are in the vdso.

Patch-set: 1