Update patch set 5
Patch Set 5:

There does not appear to be a precedent to allow a benchmark run and result to be posted in tests/. I need a continuous benchmark to (eventually) be posted in a dashboards.

Is there a mechanism to have these tests in the benchmarks, not run on normal test executions, be gated by the xml (?) and allow a command-line override to specifically run them. Did I make a mistake by adding them to full.xml/host.xml and tests/interface_test.cpp?

Patch-set: 5
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>