Update patch set 4
Patch Set 4: Code-Review+2

> Patch Set 3:
> 
> (3 comments)
> 
> > Patch Set 2:
> > 
> > (3 comments)
> > 
> > The approach looks fine, though the 65K limit is clearly an issue.
> > 
> > Presumably using the address to index into a table isn't quite legit, since the mutex can be mapped at more than one location?
> 
> I explained why I prefer 64K limit in the bug. But let me know if you think we should support more.
> pthread_mutex_t shouldn't be copied, so using address is fine, but using unordered_map can be slow, and there isn't one I can use.

What happens if you mmap a file at two different locations, allocate and initialize a mutex in one, and then access it through the other? My current reading of Posix is that's allowed, so that the same mutex can have two addresses. But I'm not sure about that interpretation.

If that were not the case, we might be able to design a custom data structure that does better than unordered_map. Unordered_map has some fairly serious constraints on asymptotic running time that we probably don't care about.

The CL looks good to me, as is.

Patch-set: 4
Reviewer: Hans Boehm <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2