Update patch set 1
Patch Set 1:

I'm fine with this to fix a P0 bug. In the slightly longer term, I think this should at least log the problem, if there's a way to do that from this level. Or we should recognize this specific case, and otherwise crash.

It seems to me that, in general, if we see this, there's a very high probability that we just destroyed the same object twice, the MUCH harder.

I'm surprised that there are no other symptoms in the motivating case here. But apparently it happens.

Patch-set: 1
Reviewer: Hans Boehm <1042828@85c56323-6fa9-3386-8a01-6480fb634889>