Update patch set 1
Patch Set 1:

I analysed some crashes reported recently during some robustness tests and I figured out that the SSE2 strlcpy is at cause.  I have isolated the issue, tried to get it fixed but it is not a simple fix, it requires a entire re-write of that optimization because the design is the cause of the issue.

We strongly suggest that this optimization gets removed.

Note that some crashes like this have been reported in the past by nobody had been able to isolate and explain the issue with enough detail to understand what was the root cause.

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Jeremy Compostella <1051574@85c56323-6fa9-3386-8a01-6480fb634889>