Update patch set 2
Patch Set 2:

> Patch Set 2: Verified+1 Autosubmit+1 Presubmit-Ready+1 Code-Review+2
> 
> we should probably add some tests (for all these <string.h> functions) that work on strings at the end of an mmaped region...

I think we could just use the XXXOverreadTest in buffer_tests.cpp. It mprotects the data at the end of the buffer to verify that the algorithm doesn't read past the end of a map.

I'll take a look and see if I can add a quick test for these. Most of the functions already have this kind of test, I think. I can also verify if this would have caught this bug.

Patch-set: 2
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>