Update patch set 2
Patch Set 2:

Thanks for the heads-up.

We don't see any functional issues for us with this change. Any adaptation that we might have to do for this should be fairly straightforward. We will test this asap but at this time there is (from our side) nothing that should prevent this from being merged.

Patch-set: 2
Reviewer: John Eckerdal <1226510@85c56323-6fa9-3386-8a01-6480fb634889>