Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> thanks. i've raised http://b/79626604 to improve our tests.
> On Wed, May 2, 2018 at 4:35 PM Compostella, Jeremy <
> jeremy.compostella@intel.com> wrote:
> 
> attachment). The principle is simple and can be extended easily.
> 
> noreply-gerritcodereview-ei93bIOc5wwulllsLDB35A@google.com>
> 
> strlcpy and srlcat."
> 
> Christopher Ferris, Mingwei Shi, Alexander Ivchenko
> 
> caught the bug you described, so it seems like we're missing something...
> (and would like to improve our test coverage!)
> 
> filters, visit settings.

Thank you for the update. We had evaluated the impact of security. According to our analysis,
In some cause, there is error in the buffer length calculation of read bytes, the read could exceed the boundary. but there is no error when copy content to the target buffers. The additional bytes is only in register and cleaned next loop. No buffer overflow in this issue. So, there is no significant security concern of the issue.

Patch-set: 2