Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> can we have the [work-in-progress from sspatil's folks] first-stage init remain static and set up /dev and /proc, and then have the second-stage init be dynamic without the need for this?

oh this topic is WIP. :)  

Anyway, the answer depends on how small the first-stage init will be with static linking. If it is not small enough (especially for recovery), it needs to be a dynamic exec as well, and we will need this.

Patch-set: 1
Reviewer: Jiyong Park <1132673@85c56323-6fa9-3386-8a01-6480fb634889>