Update patch set 2
Patch Set 2: Presubmit-Verified-1

=== Finished presubmit run: L42200000182293736 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/704835?revision=2&workplanId=L42200000182293736




TreeHugger finished with: 25 passed, 2 failed, 20 skipped.







WARNING: The tip-of-tree(4839533) for git_master/errorprone may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



FAIL: Failed to build git_master/gce_x86_phone-userdebug/4838916 (LKGB). See https://android-build.googleplex.com/builds/pending/P6165043/gce_x86_phone-userdebug/latest
FAIL: Failed to build git_master/gce_x86_phone-userdebug/4839533 (ToT). See https://android-build.googleplex.com/builds/pending/P6165038/gce_x86_phone-userdebug/latest
WARNING: The tip-of-tree(4839533) for git_master/gce_x86_phone-userdebug may be broken.
Presubmit builds based on the tip-of-tree (4839533) and the
last-known-good-build(4838916) also fail. This could be due to the
pre-existing error, a new error in this change, or both.












































WARNING: The tip-of-tree(4839533) for git_master/marlin-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.














WARNING: The tip-of-tree(4839533) for git_master/sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4839533) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4839533) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 2
Tag: autogenerated:TreeHugger:Verify-Fail
Reviewer: Treehugger Robot <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=-1