Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > For AArch64, at least, it means now both memmove and memcpy shares the same entrypoint and each call to memcpy add an extra 4 instructions and a taken branch to every call. From previous comments it seems the performance drop was acceptable, without further explanation why.
> > 
> > Since neither the standard or other libc implementation guarantee such behavior,
> 
> you're right that glibc doesn't -- and has actually moved in the opposite direction, much to Linus' annoyance -- but other important libcs do.

No libraries I know of do, especially not for AArch64 (for background: I wrote the optimized memcpy/memmove/memset for AArch64 which are widely used across libraries).

> > seems sloppy imho to push such behavior. Compiler will still conservatively auto-generate memcpy calls only for non-overlap region (unless someone adds flags to change its behavior, but it seems not a quite profitable flag since memmove is exactly the way for such) 
> 
> we're not worried about the _compiler_ making incorrect assumptions. it's humans we're trying to defend against :-)

I think you are misunderstanding Adhemerval's point: compilers will continue to use memcpy semantics. Memcpy may be inlined even if the size is not a constant, so changing the library implementation has no effect on many memcpy calls.

As a result this is not something you can change unilaterally in a library - this needs to be agreed upon between language, compiler and library developers. Otherwise it will lead to more buggy software since memcpy only has memmove semantics in some unspecified circumstances. An application may fail if a memcpy call is inlined when it previously wasn't. Is that a safety improvement?

> you'll see this a lot in bionic: from extra bounds checks that the standard doesn't mandate to things like explicit null pointer checks that give clear error messages when they abort rather than just SIGSEGV.
> 
> correctness and security are more important than tweaking spec scores.

I bet many Android vendors would disagree with slowing down such a critical function, especially since it creates new safety issues...

If increasing safety is the goal then why not abort if an incorrect overlap is detected? That would enable programmers to fix their incorrect code. Such a runtime check should be implemented in compiler so it applies to all copies and can be optimized etc.

> > and the implementation of memcpy/memmove could still be optimized (in both code size and complexity) by using a single merged function with 2 entry points (since usually memove internally branches to memcpy for non-overlap memory regions).
> 
> yes please. i'll +2 that patch :-)

It should be fixed indeed - the current version looks like a quick hack. It may be best to revert this patch, and specify the goals in more detail first. Merging memcpy and memmove implementations to simplify and share code is reasonable. Such a combined memmove+memcpy already exists in GLIBC. However changing memcpy semantics is a different matter. 

In the future it may be feasible to import memcpy and string implementations for AArch64 from Arm Optimized Routines just like the fast math functions.

> > So the question is why this behavior is preferred to Bionic?
> 
> the reason for the _behavior_ change is obvious from the philosophical point above.

Well the change is equivalent to silently ignoring serious bugs and creating a new class of bugs due to memcpy inlining.

> the reason for the effort put into the _implementation_ is that (a) that kind of thing is largely left to SoC vendors anyway and (b) this is also a small part of a much larger project we're working on for Q, where we'll switch to ifuncs and runtime selection of optimized routines rather than the build-time selection you see today. since "make memcpy safe" has been on our to-do list for years and slightly reduces the amount of work we need to do with the ifuncs, we did this first.

An ifunc would be a useful way to select between different versions indeed, including ones that do extra checking when testing/debugging. But please do not call the current change "making memcpy safe".

Patch-set: 5
CC: Wilco Dijkstra <1333281@85c56323-6fa9-3386-8a01-6480fb634889>