Update patch set 1
Patch Set 1:

I'm generally in favor of making the system call accessible. We've been talking about various potential uses for it in ART as well, though our tentative conclusion was that the non-expedited flavor is probably too slow to be useful, at least for our purposes. I presume that this is the right way to make it accessible.

But I'm very concerned about the initial use of it I found in the accompanying CL. This call is a fantastic code obfuscation tool. And probably a serious performance hazard if abused. Explicit fences are bad. This is much worse, since it's one-sided, and half of the affected code has no annotation at all. I don't believe it should ever be used without fairly elaborate justification in the code. The only use so far doesn't even have a comment. I think that absolutely has to be fixed. (In some cases, the comment may need to point to a design document. That may be necessary here.)

Patch-set: 1
Reviewer: Hans Boehm <1042828@85c56323-6fa9-3386-8a01-6480fb634889>