Update patch set 1
Patch Set 1: Code-Review+1

As for the BPF plan - while I agree that it needs an explicit comment, it was suggested by the kernel eBPF maintainer, so we believe that the other side of the fence will not go away. Basically, it boils down to the fact that membarrier(MEMBARRIER_CMD_SHARED) causes a synchronize_rcu(), and that the eBPF code uses RCU synchronization.

Hans, do you have concerns on performance implications? Will calling this on an arbitrary system_server thread cause other threads to stutter or jank? If it will just block the calling thread for as long as necessary to provide the appropriate guarantees, that's fine (assuming the grace period is of the order of milliseconds or a bit more).

Patch-set: 1
Reviewer: Lorenzo Colitti <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1