Update patch set 1
Patch Set 1:

Adding Orion. He did some earlier measurements of the membarrier system call. (I had misremembered that it was Igor, who is currently out.)

> Patch Set 1: Code-Review+1
> 
> As for the BPF plan - while I agree that it needs an explicit comment, it was suggested by the kernel eBPF maintainer, so we believe that the other side of the fence will not go away. Basically, it boils down to the fact that membarrier(MEMBARRIER_CMD_SHARED) causes a synchronize_rcu(), and that the eBPF code uses RCU synchronization.
> 
> Hans, do you have concerns on performance implications? Will calling this on an arbitrary system_server thread cause other threads to stutter or jank? If it will just block the calling thread for as long as necessary to provide the appropriate guarantees, that's fine (assuming the grace period is of the order of milliseconds or a bit more).

My concern is whether the membarrier maintainers have bought into this. I personally feel they probably shouldn't; I'd argue that adding semantic constraints to barriers that are not fundamentally related to their main functionality is not a good idea. Consider what this would do to membarrier performance on an architecture on which it could otherwise be a no-op, or has direct hardware support (variants of which have been proposed in academic papers, at least).

We have experience with the analogous hardware problem: x86 MFENCE became perhaps too heavy to use for its original purpose, presumably because it acquired too much baggage. This has caused us a bunch of unpleasantness.

The one mitigating factor is that (on some architectures) there is the "expedited" call, which hopefully wouldn't share this constraint.

Patch-set: 1
CC: Orion Hodson <1071150@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Hans Boehm <1042828@85c56323-6fa9-3386-8a01-6480fb634889>