Update patch set 1
Patch Set 1:

I see that Mathieu Desnoyers joined the upstream discussion. That's good. He was heavily involved in the last membarrier discussion, and I believe wrote a bunch of the code.

As others said, in general, any sort of memory barrier / fence needs a corresponding reordering constraint in the other thread involved in the communication. It's impossible to observe whether one threads accesses are being reordered without preventing access reordering in the other thread. The matching constraint may not be a barrier/fence itself, but generally something is needed. E.g. if the other side is written in x86 assembly, hardware ordering constraints may suffice.

The system call "one-sided" barrier means that the other side no longer needs anything to prevent hardware reordering. But it does still need a compiler barrier/fence (or assembly code).

The simplest case in which we do not want to add constraints to the membarrier system call is the one in which the hardware doesn't reorder to start with. I believe we really want to allow the no-op implementation of the system call in that case. This is rare, but such platforms do seem to reappear now and then.

Patch-set: 1
Reviewer: Hans Boehm <1042828@85c56323-6fa9-3386-8a01-6480fb634889>