Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review+2
> 
> (1 comment)
> 
> hboehm: what was the outcome of this? did the committees ever agree to cooperate? (if so, we should try to get the clang <stdatomic.h> fixed and remove the one from bionic.)

wg21.link/P0943 is still in the WG21 pipeline. The concurrency group is happy with it. The main obstacle is that it's competing with the existing deprecation of ".h" headers, and kind of with the intent of P0657, which wants to remove everything from the global namespace. I personally think these two are both misguided. (Though the main part of P0657 may be OK.) AFAICT somewhere around half the committee currently agrees with me.

Patch-set: 3
Reviewer: Hans Boehm <1042828@85c56323-6fa9-3386-8a01-6480fb634889>