Update patch set 5
Patch Set 5:

The armv8-a builds fine, however not sure which ABI clang/ndk is using for this configuration:

00011800 <fmal@@LIBC>:
   11800:       ec432b30        vmov    d16, r2, r3
   11804:       eddd2b00        vldr    d18, [sp]
   11808:       ec410b31        vmov    d17, r0, r1
   1180c:       eee12ba0        vfma.f64        d18, d17, d16
   11810:       ec510b32        vmov    r0, r1, d18 
   11814:       e12fff1e        bx      lr

000118f4 <fmaf@@LIBC>:
   118f4:       ee001a10        vmov    s0, r1  
   118f8:       ee010a10        vmov    s2, r0
   118fc:       ee022a10        vmov    s4, r2     
   11900:       eea12a00        vfma.f32        s4, s2, s0                        
   11904:       ee120a10        vmov    r0, s4          
   11908:       e12fff1e        bx      lr

The input arguments seems to be passed on general registers instead of floating point one, is this the expected one for android for armv8-a? I am asking because this differ from expected armv8 abi.

Patch-set: 5
Reviewer: Gerrit User 1331125 <1331125@85c56323-6fa9-3386-8a01-6480fb634889>