Update patch set 17
Patch Set 17:

Sorry about the time it's taking on this.

I was doing benchmarks on the new code, and I found that single thread benchmarks are exactly the same (as we would expect).

I ran all of the other threaded ones but I haven't had a chance to check the results, but it also looked like a wash. I did this for patchset 14.

I'll redo some of the threaded stuff.

By the way, I believe you need to do a rebase because you are missing the reallocarray function in malloc_common.cpp.

Patch-set: 17
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>