CTS: allow continue to use the old AIDs on devices launch with P
Some vendors may have been using the AIDs outside the oem ranges
for long that those aids need continue to be maitained for the
existing devices. We allow them countinue to use the old AID names
and values on device launch with P and before, to give them the
preiod of time to adopt the new AID scheme.

Bug: 116405200
Change-Id: Ifd0a452711e20b856adf54fb20e71b23c9636e00