Update patch set 1
Patch Set 1:

(1 comment)

What is the performance impact of this? It seems like a really expensive way to implement.

I was thinking about a way to add a much quicker way to have this kind of limit ability with jemalloc5. It wouldn't be perfect (maybe some allocations in a cache somewhere not decremented properly so you hit the limit while not really being at the limit).

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>