Update patch set 1
Patch Set 1: Code-Review+1

The only downside to this that I can see is that for bigger objects (say 0.5k/1k) the allocator consumes more pages per object. For 2k it is twice as much (but iirc we fall back on mmap on >=2k). I assume this situation does not arise often since you see the memory consumption reduced in your testing.

It would also be nice to have the results of the measurements of memory saving in commit message and additional details on how you have measured these.

Patch-set: 1
Reviewer: Gerrit User 1041673 <1041673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1