Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Could this be detected at runtime to avoid having a compile time flag?  Mainline wants to use the same system image for multiple devices, only some of which may need this.

Mainline is the thing that will be used to build APEX updates, right?

If we wanted to detect this at runtime, probably the simplest way would be to create a blacklist of bad kernel versions (i.e. uname -r). We could then just not use the vDSO at all on those kernels. I'll see if the set of bad kernel versions is small enough to make this feasible.

Patch-set: 1
Reviewer: Gerrit User 1067098 <1067098@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0