Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > > Patch Set 1:
> > > > > 
> > > > > Could this be detected at runtime to avoid having a compile time flag?  Mainline wants to use the same system image for multiple devices, only some of which may need this.
> > > > 
> > > > Mainline is the thing that will be used to build APEX updates, right?
> > > > 
> > > > If we wanted to detect this at runtime, probably the simplest way would be to create a blacklist of bad kernel versions (i.e. uname -r). We could then just not use the vDSO at all on those kernels. I'll see if the set of bad kernel versions is small enough to make this feasible.
> > > 
> > > Is the clobbering not deterministic enough to call the VDSO a few times in a static initializer and see if it gets clobbered?
> > 
> > Yeah, that's the other option. It might actually be better since I'm discovering that pretty much every QPR has its own kernel version.
> > 
> > I'll take a look at the generated assembly in the old kernels to see whether that would work too.
> 
> Some other options:
> Save x18 on any kernel earlier than the major version when the patch went upstream.

It's not quite as simple as that. The patch that broke things never landed upstream; it was applied only to our Pixel kernel branches. And the version numbers that the kernel reports generally don't correspond with whether the patch was applied due to the fact that the patch was cherry picked into the individual Pixel kernel branches. So if this were to be based on kernel versions we would need to look at the branches that corresponded to Pixel 2/3 kernel releases (including QPRs) with the patch but without the fix and blacklist those versions. This would cast a wider net than strictly necessary (e.g. if an OEM happened to base their kernel on the same major/minor/patch version as a Pixel kernel they would also receive the workaround), but maybe that's acceptable.

> Always save x18, might be cheaper than an extra conditional?

I'd prefer not to since this could leak the SCS pointer. Also the conditional would be in the initialization code so it shouldn't affect performance of the time-related libc functions on the "golden path" (i.e. new kernel).

Patch-set: 1
Reviewer: Gerrit User 1067098 <1067098@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0