Update patch set 2
Patch Set 2:

You should abandon this cl and do a real cherry-pick (choose the download option, copy the cherry-pick option, add a -x to the command and run it in aosp bionic). Also update the description to add the line Merged-In:, with a copy of the Commit-Id after the Merged-In: directive.

Patch-set: 2
CC: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>