Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm currently skipping x18 in the tlsdesc resolver function. I *think* that's the correct behavior with respect to Shadow Call Stack, but I figured I'd add pcc to the CL.
> 
> I noticed a comment in the arm64 setjmp.S: "We only store the low bits of x18 to avoid leaking the shadow call stack address into memory.")
> 
> The tlsdesc resolver needs to call a C++ function, __tls_get_addr, and save/restore any (volatile) register that __tls_get_addr might have modified, with a few exceptions like x0 (and lr, and AFAICT, the nzcv condition codes).

FWIW, LLVM models the TLSDESC call as clobbering only X0, X1 and LR:
http://llvm-cs.pcc.me.uk/lib/Target/AArch64/AArch64InstrInfo.td#1639
I think this means that NZCV needs to be preserved as well.

> It doesn't need to preserve callee-save registers. My impression is that a function call won't modify the x18 register if it's being used for the SCS register.

Correct. Even if the runtime code was compiled with SCS, the register should be restored to its original value upon returning from the function, similar to a stack pointer.

> The tlsdesc resolver currently only calls into linker code that should be compiled to reserve x18. Eventually, I think it could call into a sanitizer hook (e.g. the registrations suggested in https://sourceware.org/glibc/wiki/ThreadPropertiesAPI), but those should also be reserving x18.

Correct. Since any sanitizer hooks that we might create would be a new API, we can expect the hooks to be compiled with a new compiler that will preserve x18.

Patch-set: 1
Reviewer: Gerrit User 1067098 <1067098@85c56323-6fa9-3386-8a01-6480fb634889>