Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> Looks good to me.
> 
> I was wondering when the ifunc resolvers should run. I assume it needs to run before apply_gnu_relro because it could relocate something in RELRO.
> 
> If we wanted to match dynamic executables, I think it'd go between __libc_init_AT_SECURE and layout_static_tls.
> 
> AFAIK, we're not using ifuncs in libc.a yet, so running them as late as possible probably makes sense. I think that's what this CL is doing.

It does look like it would need to be moved earlier if we want libc.a to use ifuncs. I tried replacing libc_common with libc_common_shared in libc.a and saw crashes on startup in bionic-unit-tests-static because __libc_init_common calls __register_atfork calls malloc, which can basically call anything. In the place you suggested I saw failing tests in bionic-unit-tests-static:

[   FAIL   ] 4 tests failed, listed below:
[   FAIL   ] stack_protector_DeathTest.modify_stack_protector
[   FAIL   ] setjmp.setjmp_cookie
[   FAIL   ] stack_protector.same_guard_per_thread
[   FAIL   ] stack_protector.global_guard

I was able to get the tests to pass by moving the call right after __libc_init_main_thread_early. I'm not sure if we want to move it that early, though. It might be a good idea to move it as late as possible and compile the early startup code with -ffreestanding to prevent the compiler from inserting calls to memset et al.

Patch-set: 2