Update patch set 4
Patch Set 4:

Whoops, sorry for the spam there. The VTS change conflicted, but fortunately it was because the bionic tests were removed from vts, so I've just dropped it - all this needs now is a fresh +2 for the change with the added test code.

Patch-set: 4
Reviewer: Gerrit User 1021004 <1021004@85c56323-6fa9-3386-8a01-6480fb634889>