Suppress warnings on pthread_create
This commit suppress warnings on pthread_create because clang-r353983
would check the function declaration against the built-in function
defined in llvm/tools/clang/include/clang/Basic/Builtins.def and find a
mismatch.

Note: This is only found by versioner because these files are not system
headers from the perspective of versioner.  This warning is ignored in
normal builds because bionic headers are system headers in normal
builds.

Bug: 126457671
Test: lunch walleye-userdebug && make
Change-Id: I3f05ba19861f1b9db55c7c55c4496a845802e831