Update patch set 1
Patch Set 1: Code-Review+2

I agree that this is the way things work, and I think it's great to document that.

I would also explicitly state the implication of that:

For all intents and purposes futex waits can wake up spuriously, even if the kernel doesn't let that happen. The problem is that we can do futex wakes on arbitrary addresses that used to belong to a mutex, but may have since been reallocated to anything else.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2